Finishing up - hopefully!

+++ libraries/Time/Time.cpp
@@ -160,25 +160,25 @@ time64_t makeTime(tmElements_t &tm){
   time64_t seconds; // was uint32_t seconds **** fix for y2038 bug
  
   // seconds from 1970 till 1 jan 00:00:00 of the given year
-  seconds= (unsigned long long)tm.Year*(SECS_PER_DAY * 365);
+  seconds= tm.Year*(SECS_PER_DAY * 365);
   for (i = 0; i < tm.Year; i++) {
     if (LEAP_YEAR(i)) {
-      seconds += (unsigned long long)SECS_PER_DAY;   // add extra days for leap years
+      seconds += SECS_PER_DAY;   // add extra days for leap years
     }
   }
   
   // add days for this year, months start from 1
   for (i = 1; i < tm.Month; i++) {
     if ( (i == 2) && LEAP_YEAR(tm.Year)) { 
-      seconds += (unsigned long long)SECS_PER_DAY * 29;
+      seconds += SECS_PER_DAY * 29;
     } else {
-      seconds += (unsigned long long)SECS_PER_DAY * monthDays[i-1];  //monthDay array starts from 0
+      seconds += SECS_PER_DAY * monthDays[i-1];  //monthDay array starts from 0
     }
   }
-  seconds+= (unsigned long long)(tm.Day-1) * SECS_PER_DAY;
-  seconds+= (unsigned long long)tm.Hour * SECS_PER_HOUR;
-  seconds+= (unsigned long long)tm.Minute * SECS_PER_MIN;
-  seconds+= (unsigned long long)tm.Second;
+  seconds+= (tm.Day-1) * SECS_PER_DAY;
+  seconds+= tm.Hour * SECS_PER_HOUR;
+  seconds+= tm.Minute * SECS_PER_MIN;
+  seconds+= tm.Second;
 
   //t64.time = seconds; I2CSerial.printf("\n\nmakeTime() seconds=%lx %lx\n", t64.words[1], t64.words[0]);
